+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add BILRESA control mode changing endpoint.
 - Fix Hub types.
 - Fix Light types.
+- Fix Scene types.
 
 ## [1.7.0](https://github.com/lpgera/dirigera/compare/v1.6.0...v1.7.0) - 2025-12-27